+++ Sources/View/Channel/MessageCell/MessageCellParams/SBUUserMessageCellParams.swift
@@ -14,11 +14,14 @@ public class SBUUserMessageCellParams: SBUBaseMessageCellParams {
     }
     public let useReaction: Bool
     public let withTextView: Bool
+    /// The boolead value to indicates that the message cell should hide quick reply. If it's `true`, never show the quick reply view even the `UserMessage/data` has the reply option values.
+    public let shouldHideQuickReply: Bool
     
-    public init(message: UserMessage, hideDateView: Bool, useMessagePosition: Bool, groupPosition: MessageGroupPosition = .none, receiptState: SBUMessageReceiptState = .none, useReaction: Bool = false, withTextView: Bool, isThreadMessage: Bool = false, joinedAt: Int64 = 0) {
+    public init(message: UserMessage, hideDateView: Bool, useMessagePosition: Bool, groupPosition: MessageGroupPosition = .none, receiptState: SBUMessageReceiptState = .none, useReaction: Bool = false, withTextView: Bool, isThreadMessage: Bool = false, joinedAt: Int64 = 0, shouldHideQuickReply: Bool = true) {
         
         self.useReaction = useReaction
         self.withTextView = withTextView
+        self.shouldHideQuickReply = shouldHideQuickReply
         
         var messagePosition: MessagePosition = .left
         if useMessagePosition {