+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+### v3.7.0-beta.0 (Jul 17, 2023)
+* Added quick reply features
+    * Added `SBUQuickReplyView`, `SBUQuickRepViewParams` and `SBUQuickReplyViewDelegate`
+    * Added `SBUQuickReplyOptionView`, `SBUQuickReplyOptionViewDelegate`
+    * Added `quickReplyView` and `quickReplySelectHandler` to `SBUUserMessageCell`
+    * Added `groupChannelModule(_:didSelectQuickReplyOption:)` to `SBUGroupChannelModuleListDelegate`
+* Added card list features
+    * Added `cardViewParamsCollectionBuilder` to `SBUGlobalCustomParams`
+    * Added `SBUCardListData<CardData: Codable>`
+    * Added `SBUCardListView` and `SBUCardListViewParams`
+    * Added `SBUCardView` and `SBUCardViewParams`
+    * Added `cardListView` to `SBUUserMessageCell`
+
 ### v3.6.2 (Jul 14, 2023) with Chat SDK **v4.9.5**
 * Changed file cache key policy.
 * Improved the issue of exposing empty images when message status updating from pending to succeed.